+++ beacon-chain/core/blocks/attestation_test.go
@@ -1,0 +1,1 @@
+package blocks
