+++ beacon-chain/core/blocks/attester_slashing.go
@@ -1,0 +1,1 @@
+package blocks
