+++ beacon-chain/core/blocks/deposit.go
@@ -1,0 +1,1 @@
+package blocks
