+++ beacon-chain/core/blocks/eth1_data.go
@@ -1,0 +1,1 @@
+package blocks
