+++ beacon-chain/core/blocks/exit.go
@@ -1,0 +1,1 @@
+package blocks
