+++ beacon-chain/core/blocks/exit_test.go
@@ -1,0 +1,1 @@
+package blocks
