+++ beacon-chain/core/blocks/genesis.go
