+++ beacon-chain/core/blocks/genesis_test.go
