+++ beacon-chain/core/blocks/header.go
@@ -1,0 +1,1 @@
+package blocks
