+++ beacon-chain/core/blocks/header_test.go
@@ -1,0 +1,1 @@
+package blocks
