+++ beacon-chain/core/blocks/log.go
@@ -1,0 +1,1 @@
+package blocks
