+++ beacon-chain/core/blocks/proposer_slashing_test.go
@@ -1,0 +1,1 @@
+package blocks
