+++ beacon-chain/core/blocks/randao.go
@@ -1,0 +1,1 @@
+package blocks
