+++ beacon-chain/core/blocks/randao_test.go
@@ -1,0 +1,1 @@
+package blocks
